api/requirements.txt
Django
psycopg2
djangorestframework
django-cors-headers
django-rest-swagger
django-rest-knox
pytest
